Bug 1159561 - Pre: move DatabaseError into its own file.
